reorganize for config
