+++ config.ru
@@ -4,4 +4,4 @@ $LOAD_PATH << 'lib'
 
 require 'make_it_rain'
 use Rack::CommonLogger, STDOUT
-run MakeItRain.new
+run MakeItRain::Wrapper
